rename pkg to elemental + fix basic routing + add simple translations
