+++ pkg/elemental/babel.config.js
