+++ pkg/elemental/pages/MachineInventories.vue
@@ -1,0 +1,13 @@
+<script>
+export default { name: 'MachineInventories' };
+</script>
+
+<template>
+  <div>
+    <h1>MachineInventories!!!</h1>
+  </div>
+</template>
+
+<style lang="scss" scoped>
+
+</style>
