+++ pkg/elemental/types.js
@@ -1,4 +1,4 @@
-export const ELEMENTAL_PRODUCT_NAME = 'os-elemental';
+export const ELEMENTAL_PRODUCT_NAME = 'elemental';
 
 export const ELEMENTAL_TYPES = {
   DASHBOARD:           'dashboard',
